Split MapboxStatic.swift into multiple Swift files
